+++ CHANGELOG.md
@@ -5,12 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.0.2 - 2018-02-17
+
+### Added in 0.0.2
+
+- Added KvLang language server
+- Improved syntaxes of KvLang
+- Corrected few snippets
+- Added basic parser for the KvLang using language server. Parsing errors are displayed during saving file
+- Added settings for KvLang extension
+
 ## 0.0.1 - 2018-01-20
 
-### Added
+### Added in 0.0.1
 
 - Initial release
-- Added basic syntaxes of kvlang
+- Added basic syntaxes of KvLang
 - Added basic language configuration
 - Added kivy language snippets
 - Added basic snippets of the uix widget names